Refactor Options. Move toHtml() to PegdownDoclet.
