+++ doclet/build.gradle
@@ -8,6 +8,8 @@ dependencies {
     compile libs.plantuml
     compileOnly libs.jdkTools
 
+    compile group: 'org.codehaus.groovy', name: 'groovy-all', version: '2.4.6'
+
     testCompile group: 'org.jsoup', name: 'jsoup', version: '1.8.3'
 }
